fix issue of missing Paths_<package_name>
Paths_hello is generated by cabal during the configure stage. With our
override, this step gets skipped and we get a compilation error. So we
need the override to trigger the configuration and build steps again.

I'm not using substitute because cabal was complaining of unrecognized
arguments, presubably because arguments meant for `cabal test` were
being passed to `cabal configure` and `cabal build`. Now I'm a bit
concerned about throwing away the user's arguments...