complete rewrite.
I added some logging code, and that allowed me to see that I was
misunderstanding the relationship between cabal, Setup.hs, and `args`.
I added a comment explaining that, and I rewrote the solution. It's much
simpler now, it supports all the previous corner cases, *and* it doesn't
lose any user arguments!